Move server creation
